[Refactor] 결과 데이터 파일 분리 #33
