+++ src/app/landing/data.ts
@@ -1,0 +1,55 @@
+interface resultType {
+  [result: string]: { productNumber: number; content: string };
+}
+
+export const question = [
+  { title: "평소 주량이 어떻게 되세요?", type: "alchol", A: "half", B: "1bottle", C: "2bottle" },
+  { title: "원하시는 당도를 선택해주세요!", type: "sweet", A: 0, B: 50, C: 100 },
+  {
+    title: "술과 언제 함께 하실건가요?",
+    type: "situation",
+    A: "summer",
+    B: "rainy",
+    C: "romantic",
+    D: "friend",
+  },
+];
+
+export const result: resultType = {
+  AAA: { productNumber: 123, content: "AAA 술 추천 문구~~~~" },
+  AAB: { productNumber: 123, content: "AAB 술 추천 문구~~~~" },
+  AAC: { productNumber: 123, content: "AAC 술 추천 문구~~~~" },
+  AAD: { productNumber: 123, content: "AAD 술 추천 문구~~~~" },
+  ABA: { productNumber: 123, content: "ABA 술 추천 문구~~~~" },
+  ABB: { productNumber: 123, content: "ABB 술 추천 문구~~~~" },
+  ABC: { productNumber: 123, content: "ABC 술 추천 문구~~~~" },
+  ABD: { productNumber: 123, content: "ABD 술 추천 문구~~~~" },
+  ACA: { productNumber: 123, content: "ACA 술 추천 문구~~~~" },
+  ACB: { productNumber: 123, content: "ACB 술 추천 문구~~~~" },
+  ACC: { productNumber: 123, content: "ACC 술 추천 문구~~~~" },
+  ACD: { productNumber: 123, content: "ACD 술 추천 문구~~~~" },
+  BAA: { productNumber: 123, content: "BAA 술 추천 문구~~~~" },
+  BAB: { productNumber: 123, content: "BAB 술 추천 문구~~~~" },
+  BAC: { productNumber: 123, content: "BAC 술 추천 문구~~~~" },
+  BAD: { productNumber: 123, content: "BAD 술 추천 문구~~~~" },
+  BBA: { productNumber: 123, content: "BBA 술 추천 문구~~~~" },
+  BBB: { productNumber: 123, content: "BBB 술 추천 문구~~~~" },
+  BBC: { productNumber: 123, content: "BBC 술 추천 문구~~~~" },
+  BBD: { productNumber: 123, content: "BBD 술 추천 문구~~~~" },
+  BCA: { productNumber: 123, content: "BCA 술 추천 문구~~~~" },
+  BCB: { productNumber: 123, content: "BCB 술 추천 문구~~~~" },
+  BCC: { productNumber: 123, content: "BCC 술 추천 문구~~~~" },
+  BCD: { productNumber: 123, content: "BCD 술 추천 문구~~~~" },
+  CAA: { productNumber: 123, content: "CAA 술 추천 문구~~~~" },
+  CAB: { productNumber: 123, content: "CAB 술 추천 문구~~~~" },
+  CAC: { productNumber: 123, content: "CAC 술 추천 문구~~~~" },
+  CAD: { productNumber: 123, content: "CAD 술 추천 문구~~~~" },
+  CBA: { productNumber: 123, content: "CBA 술 추천 문구~~~~" },
+  CBB: { productNumber: 123, content: "CBB 술 추천 문구~~~~" },
+  CBC: { productNumber: 123, content: "CBC 술 추천 문구~~~~" },
+  CBD: { productNumber: 123, content: "CBD 술 추천 문구~~~~" },
+  CCA: { productNumber: 123, content: "CCA 술 추천 문구~~~~" },
+  CCB: { productNumber: 123, content: "CCB 술 추천 문구~~~~" },
+  CCC: { productNumber: 123, content: "CCC 술 추천 문구~~~~" },
+  CCD: { productNumber: 123, content: "CCD 술 추천 문구~~~~" },
+};
